feat: domain

+++ packages/types/src/primitive_amount/error.rs
@@ -1,9 +1,11 @@
+use rust_decimal::Decimal;
 #[derive(Debug, thiserror::Error)]
+
 pub enum PrimitiveError {
     #[error("Value cannot be negative: {0}")]
-    Negative(f64),
+    Negative(Decimal),
     #[error("Value must be positive: {0}")]
-    NonPositive(f64),
+    NonPositive(Decimal),
     #[error("Value cannot be NaN")]
     NotANumber,
     #[error("Value cannot be infinite")]
